Allow multiple files in metrics endpoint
This commit adds support for calculating the metrics for multiple asreview files at the same time. If only one file is provided, the output is the same as before. If multiple files are provided, the filepath is printed before printing the stats. Also, if an output location is specified, the output will be a dictionary with the filepaths as keys and the dictionary containing the stats as value.